+++ bugsnag/asgi.py
@@ -4,7 +4,7 @@ from typing import Any, List, Dict, Union, Optional
 import bugsnag
 from bugsnag.breadcrumbs import BreadcrumbType
 from bugsnag.legacy import _auto_leave_breadcrumb
-from bugsnag.utils import remove_query_from_url
+from bugsnag.utils import remove_query_from_url, sanitize_url
 
 __all__ = ('BugsnagMiddleware',)
 
@@ -103,7 +103,10 @@ class BugsnagMiddleware:
                         request[prop.metadata_key] = scope[item]
                         break
 
-            request['url'] = parse_url(request, server)
+            request['url'] = sanitize_url(
+                parse_url(request, server),
+                event.config
+            )
 
             event.add_tab("request", request)
             if bugsnag.configure().send_environment: